+++ src/features/chat/hook/use-leave-chat.ts
@@ -1,0 +1,1 @@
+import { useMutation } from '@tanstack/react-query';
