+++ src/features/chat/ui/chat-message-action-box.tsx
@@ -5,7 +5,7 @@ import useUploadPostImageMutation from '@/features/posts/api/use-upload-post-ima
 import type { ValueOf } from '@/shared/type/types';
 import ImageUploader from '@/shared/ui/image-uploader';
 import { type ReactNode } from 'react';
-import useSendMessage from '../hook/useSendMessage';
+import useSendMessage from '../hook/use-send-message';
 import { createChatImagesMessage } from '../lib/create-chat-images-message';
 import { HairConsultationChatMessageTypeEnum } from '../type/hair-consultation-chat-message-type';
 import type { UserHairConsultationChatChannelType } from '../type/user-hair-consultation-chat-channel-type';
